rename HfDatasets to ConcatDataset
