+++ mmchat/datasets/__init__.py
@@ -1,3 +1,4 @@
+from .concat_dataset import ConcatDataset
 from .huggingface import process_hf_dataset
 
-__all__ = ['process_hf_dataset']
+__all__ = ['process_hf_dataset', 'ConcatDataset']
